+++ app/src/main/java/ch/hevs/androidproject644/js/workingtime/MainActivity.java
@@ -99,7 +99,7 @@ public class MainActivity extends AppCompatActivity implements NavigationView.On
 
     }
 
-    private void setListener(final int taskId) {
+    private void setListener(final long taskId) {
         _lvtasks.setOnItemClickListener(new AdapterView.OnItemClickListener() {
             @Override
             public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
