+++ app/src/main/java/ch/hevs/androidproject644/js/workingtime/model/Company.java
@@ -8,12 +8,12 @@ import android.os.Parcelable;
  */
 
 public class Company implements Parcelable {
-    private int _id;
+    private long _id;
     private String _name;
     private boolean _active;
     private long _duration = 0;
 
-    public Company(int id, String name, boolean active) {
+    public Company(long id, String name, boolean active) {
         this._id = id;
         this._name = name;
         this._active = active;
@@ -22,7 +22,7 @@ public class Company implements Parcelable {
     public Company() {    }
 
     // SETTERS
-    public void set_id(int id) {
+    public void set_id(long id) {
         this._id = id;
     }
     public void set_name(String name) {
@@ -36,7 +36,7 @@ public class Company implements Parcelable {
     }
 
     // GETTERS
-    public int get_id() {
+    public long get_id() {
         return _id;
     }
     public String get_name() {
@@ -66,7 +66,7 @@ public class Company implements Parcelable {
     };
 
     protected Company(Parcel in) {
-        this._id = in.readInt();
+        this._id = in.readLong();
         this._name = in.readString();
 
         boolean[] myBooleanArr = new boolean[1];
@@ -78,7 +78,7 @@ public class Company implements Parcelable {
 
     @Override
     public void writeToParcel(Parcel dest, int flags) {
-        dest.writeInt(_id);
+        dest.writeLong(_id);
         dest.writeString(_name.toString());
 
         boolean[] myBooleanArr = {_active}; //= new boolean[1];
