+++ app/src/main/java/ch/hevs/androidproject644/js/workingtime/model/Time.java
@@ -19,20 +19,20 @@ import ch.hevs.androidproject644.js.workingtime.Controler.C_Time;
  */
 
 public class Time implements Parcelable {
-    private int _id;
+    private long _id;
     private Calendar _start;
     private Calendar _stop;
     private long _duration;
-    private int taskId;
+    private long taskId;
 
 
     protected Time(Parcel in) {
-        this._id = in.readInt();
-        this._duration = in.readInt();
+        this._id = in.readLong();
+        this._duration = in.readLong();
         this._start = (Calendar) in.readValue(getClass().getClassLoader());
         this._stop = (Calendar) in.readValue(getClass().getClassLoader());
         //this._task = in.readParcelable(Task.class.getClassLoader());
-        this.taskId = in.readInt();
+        this.taskId = in.readLong();
 
     }
 
@@ -47,7 +47,7 @@ public class Time implements Parcelable {
         set_start(C_Time.getCurrentTimeInSecond());
         this._task = _task;
     }*/
-    public Time (int taskId){
+    public Time (long taskId){
         set_start(C_Time.getCurrentTimeInSecond());
         this.taskId=taskId;
     }
@@ -64,12 +64,12 @@ public class Time implements Parcelable {
 
     @Override
     public void writeToParcel(Parcel dest, int flags) {
-        dest.writeInt(_id);
+        dest.writeLong(_id);
         dest.writeValue(_start);
         dest.writeValue(_stop);
         dest.writeLong(_duration);
         //dest.writeParcelable(_task,flags);
-        dest.writeInt(taskId);
+        dest.writeLong(taskId);
     }
 
     public void cal_duration() {
@@ -97,11 +97,11 @@ public class Time implements Parcelable {
         this._stop = _stop;
     }
 
-    public void set_id(int _id) {
+    public void set_id(long _id) {
         this._id = _id;
     }
 
-    public void set_task(int taskid) {
+    public void set_task(long taskid) {
         this.taskId = taskid;
     }
 
@@ -109,11 +109,11 @@ public class Time implements Parcelable {
         this._duration = _duration;
     }
 
-    public int get_id() {
+    public long get_id() {
         return _id;
     }
 
-    public int  get_task() {
+    public long  get_task() {
         return taskId;
     }
